Update demo recipe and alt text tools module
