--- modules/alt_text_tools/alt_text_tools
@@ -1,1 +1,0 @@
-/var/www/html/modules/alt_text_tools
