--- recipes/fetch_mcp_client_demo/recipe.yml
@@ -1,6 +1,0 @@
-name: 'Fetch MCP Client Demo'
-description: 'Demo to setup the MCP Client module and use the open fetch MCP server.'
-type: 'AI Demo'
-
-install:
-  - mcp_client
